# api/openapi.yaml
openapi: 3.1.0
info:
  title: buffgo local control API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/nodes:
    get:
      operationId: listNodes
      responses:
        "200":
          description: Credential-safe access nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeCreate"
      responses:
        "201":
          description: Node created in validating state; proxy credential is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    get:
      operationId: getNode
      responses:
        "200":
          description: Credential-safe node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}/name:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    post:
      operationId: renameNode
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeRename"
      responses:
        "200":
          description: Name updated; connection and assignment stay unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}/connection:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    post:
      operationId: replaceNodeConnection
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeConnectionReplace"
      responses:
        "200":
          description: Connection replaced; node returns to validating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}/exit:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    post:
      operationId: recordNodeExit
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeExit"
      responses:
        "200":
          description: Manual exit evidence recorded; node becomes available when valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}/assign-game:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    post:
      operationId: assignNodeGame
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeAssignGame"
      responses:
        "200":
          description: Game quota assigned or cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}/assign-side:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    post:
      operationId: assignNodeSide
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeAssignSide"
      responses:
        "200":
          description: Game-platform-side assignment set or cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes/{id}/delete:
    parameters:
      - $ref: "#/components/parameters/NodeID"
    post:
      operationId: deleteNode
      security:
        - controlSession: []
          csrfToken: []
      responses:
        "200":
          description: Node deleted
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    NodeID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  securitySchemes:
    controlSession:
      type: apiKey
      in: cookie
      name: buffgo_control_session
    csrfToken:
      type: apiKey
      in: header
      name: X-Buffgo-CSRF
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - invalid_host
            - method_not_allowed
            - origin_rejected
            - origin_required
            - json_required
            - csrf_rejected
            - internal_error
            - invalid_account_id
            - account_not_found
            - account_occupied
            - account_in_use
            - account_revision_conflict
            - session_required
            - session_too_large
            - invalid_json
            - body_not_allowed
            - resource_storage_unavailable
            - resource_integrity
            - resource_error
            - not_found
            - invalid_account
            - account_conflict
            - invalid_node_id
            - node_not_found
            - node_occupied
            - node_in_use
            - node_revision_conflict
            - node_assignment_conflict
            - node_name_conflict
            - platform_region_mismatch
            - invalid_node
            - invalid_exit_address
            - invalid_combination_id
            - combination_not_found
            - combination_occupied
            - combination_incompatible
            - combination_conflict
            - invalid_combination
            - resource_in_use
            - credential_required
            - credential_too_large
            - invalid_provider_id
            - provider_not_found
            - provider_revision_conflict
            - provider_conflict
            - invalid_provider
            - invalid_target_id
            - target_not_found
            - collection_not_found
            - collection_conflict
            - target_not_stopped
            - collection_in_use
            - invalid_collection
            - collection_storage_unavailable
            - collection_error
            - invalid_run_id
            - invalid_page_sequence
            - page_payload_not_found
            - invalid_limit
            - invalid_offset
            - invalid_price_range
            - invalid_appid
            - invalid_quotes
...
    Node:
      type: object
      additionalProperties: false
      required: [id, name, kind, region, egress_mode, state, egress_revision, assignment_revision, sides, has_proxy_credential]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        kind: {type: string, enum: [direct, proxy]}
        region: {type: string, enum: [domestic, foreign, hongkong]}
        egress_mode: {type: string, enum: [static, sticky]}
        state: {type: string, enum: [validating, available, unavailable]}
        egress_revision: {type: integer, format: int64, minimum: 1}
        assignment_revision: {type: integer, format: int64, minimum: 1}
        appid: {type: integer, format: int64, minimum: 1}
        sides:
          type: array
          items:
            $ref: "#/components/schemas/NodeSide"
        has_proxy_credential: {type: boolean}
        sticky_session_valid_until: {type: string, format: date-time}
        exit:
          $ref: "#/components/schemas/NodeExitView"
    NodeSide:
      type: object
      additionalProperties: false
      required: [platform, side]
      properties:
        platform: {type: string}
        side: {type: string, enum: [bid, ask]}
    NodeExitView:
      type: object
      additionalProperties: false
      required: [address, verified_at, valid_until]
      properties:
        address: {type: string}
        verified_at: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
    NodeCreate:
      type: object
      additionalProperties: false
      required: [name, kind, region, egress_mode]
      properties:
        name: {type: string}
        kind: {type: string, enum: [direct, proxy]}
        region: {type: string, enum: [domestic, foreign, hongkong]}
        egress_mode: {type: string, enum: [static, sticky]}
        sticky_session_valid_until: {type: string, format: date-time}
        proxy_credential: {type: string, writeOnly: true}
    NodeRename:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 128}
    NodeConnectionReplace:
      type: object
      additionalProperties: false
      required: [expected_egress_revision, kind, region, egress_mode]
      properties:
        expected_egress_revision: {type: integer, format: int64, minimum: 1}
        kind: {type: string, enum: [direct, proxy]}
        region: {type: string, enum: [domestic, foreign, hongkong]}
        egress_mode: {type: string, enum: [static, sticky]}
        sticky_session_valid_until: {type: string, format: date-time}
        proxy_credential: {type: string, writeOnly: true}
    NodeExit:
      type: object
      additionalProperties: false
      required: [expected_egress_revision, address, valid_until]
      properties:
        expected_egress_revision: {type: integer, format: int64, minimum: 1}
        address: {type: string}
        valid_until: {type: string, format: date-time}
    NodeAssignGame:
      type: object
      additionalProperties: false
      required: [expected_assignment_revision]
      properties:
        expected_assignment_revision: {type: integer, format: int64, minimum: 1}
        appid: {type: integer, format: int64, minimum: 0}
    NodeAssignSide:
      type: object
      additionalProperties: false
      required: [expected_assignment_revision, platform]
      properties:
        expected_assignment_revision: {type: integer, format: int64, minimum: 1}
        platform: {type: string}
        side: {type: string, enum: [bid, ask]}
...
  responses:
    Error:
      description: Request rejected without reflecting request data or credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"